+++ emojis.py
